refactor token metadata utils
