+++ token-metadata/program/Cargo.toml
@@ -18,14 +18,15 @@ num-derive = "0.3"
 arrayref = "0.3.6"
 num-traits = "0.2"
 solana-program = "1.10"
-mpl-token-vault = { version = "0.1.0", features = [ "no-entrypoint" ] }
-spl-token = { version = "3.2.0", features = [ "no-entrypoint" ] }
+mpl-token-vault = { version = "0.1.0", features = ["no-entrypoint"] }
+spl-token = { version = "3.2.0", features = ["no-entrypoint"] }
 spl-associated-token-account = { version = "~1.0.5", features = ["no-entrypoint"] }
 thiserror = "1.0"
 borsh = "0.9.2"
 shank = { version = "0.0.9" }
 serde = { version = "1.0.136", optional = true }
 serde_with = { version = "1.12.0", optional = true }
+mpl-utils = "0.0.3"
 
 [dev-dependencies]
 solana-sdk = "1.10"
