+++ token-metadata/program/src/assertions/collection.rs
@@ -1,13 +1,13 @@
 use solana_program::{account_info::AccountInfo, program_error::ProgramError, pubkey::Pubkey};
 
 use crate::{
+    assertions::assert_derivation,
     error::MetadataError,
     pda::find_collection_authority_account,
     state::{
         Collection, CollectionAuthorityRecord, MasterEditionV2, Metadata, TokenMetadataAccount,
         TokenStandard, EDITION, PREFIX,
     },
-    utils::assert_derivation,
 };
 
 pub fn assert_collection_update_is_valid(
