+++ token-metadata/program/src/assertions/mod.rs
@@ -1,2 +1,7 @@
+pub use misc::*;
+
 pub mod collection;
+pub mod edition;
+pub mod metadata;
+mod misc;
 pub mod uses;
