+++ token-metadata/program/src/assertions/uses.rs
@@ -1,10 +1,10 @@
 use solana_program::{account_info::AccountInfo, program_error::ProgramError, pubkey::Pubkey};
 
 use crate::{
+    assertions::assert_derivation,
     error::MetadataError,
     pda,
     state::{UseAuthorityRecord, UseMethod, Uses, PREFIX, USER},
-    utils::assert_derivation,
 };
 
 pub fn assert_valid_use(
