+++ token-metadata/program/src/utils.rs
@@ -1,391 +1,36 @@
-use crate::{
-    assertions::{collection::assert_collection_update_is_valid, uses::assert_valid_use},
-    deser::clean_write_metadata,
-    error::MetadataError,
-    pda::find_master_edition_account,
-    state::{
-        get_reservation_list, CollectionDetails, Creator, Data, DataV2, Edition, EditionMarker,
-        Key, MasterEditionV1, MasterEditionV2, Metadata, TokenMetadataAccount, TokenStandard, Uses,
-        EDITION, EDITION_MARKER_BIT_SIZE, MAX_CREATOR_LIMIT, MAX_EDITION_LEN,
-        MAX_EDITION_MARKER_SIZE, MAX_MASTER_EDITION_LEN, MAX_METADATA_LEN, MAX_NAME_LENGTH,
-        MAX_SYMBOL_LENGTH, MAX_URI_LENGTH, PREFIX,
-    },
-};
-use arrayref::{array_mut_ref, array_ref, array_refs, mut_array_refs};
+use arrayref::{array_mut_ref, array_ref, mut_array_refs};
 use borsh::BorshSerialize;
+use mpl_utils::{
+    assert_signer, create_or_allocate_account_raw,
+    token::{get_mint_authority, get_mint_decimals, get_mint_freeze_authority, get_mint_supply},
+};
 use solana_program::{
-    account_info::AccountInfo,
-    borsh::try_from_slice_unchecked,
-    entrypoint::ProgramResult,
-    msg,
-    program::{invoke, invoke_signed},
-    program_error::ProgramError,
-    program_option::COption,
-    program_pack::{IsInitialized, Pack},
-    pubkey::Pubkey,
-    system_instruction,
-    sysvar::{rent::Rent, Sysvar},
+    account_info::AccountInfo, borsh::try_from_slice_unchecked, entrypoint::ProgramResult, msg,
+    program::invoke_signed, program_error::ProgramError, program_option::COption, pubkey::Pubkey,
 };
 use spl_token::{
     instruction::{set_authority, AuthorityType},
-    state::{Account, Mint},
+    state::Account,
 };
-use std::{collections::HashMap, convert::TryInto};
 
-pub fn assert_data_valid(
-    data: &Data,
-    update_authority: &Pubkey,
-    existing_metadata: &Metadata,
-    allow_direct_creator_writes: bool,
-    update_authority_is_signer: bool,
-) -> ProgramResult {
-    if data.name.len() > MAX_NAME_LENGTH {
-        return Err(MetadataError::NameTooLong.into());
-    }
-
-    if data.symbol.len() > MAX_SYMBOL_LENGTH {
-        return Err(MetadataError::SymbolTooLong.into());
-    }
-
-    if data.uri.len() > MAX_URI_LENGTH {
-        return Err(MetadataError::UriTooLong.into());
-    }
-
-    if data.seller_fee_basis_points > 10000 {
-        return Err(MetadataError::InvalidBasisPoints.into());
-    }
-
-    if let Some(creators) = &data.creators {
-        if creators.len() > MAX_CREATOR_LIMIT {
-            return Err(MetadataError::CreatorsTooLong.into());
-        }
-
-        if creators.is_empty() {
-            return Err(MetadataError::CreatorsMustBeAtleastOne.into());
-        }
-
-        // Store caller-supplied creator's array into a hashmap for direct lookup.
-        let new_creators_map: HashMap<&Pubkey, &Creator> =
-            creators.iter().map(|c| (&c.address, c)).collect();
-
-        // Do not allow duplicate entries in the creator's array.
-        if new_creators_map.len() != creators.len() {
-            return Err(MetadataError::DuplicateCreatorAddress.into());
-        }
-
-        // If there is an existing creator's array, store this in a hashmap as well.
-        let existing_creators_map: Option<HashMap<&Pubkey, &Creator>> = existing_metadata
-            .data
-            .creators
-            .as_ref()
-            .map(|existing_creators| existing_creators.iter().map(|c| (&c.address, c)).collect());
-
-        // Loop over new creator's map.
-        let mut share_total: u8 = 0;
-        for (address, creator) in &new_creators_map {
-            // Add up creator shares.  After looping through all creators, will
-            // verify it adds up to 100%.
-            share_total = share_total
-                .checked_add(creator.share)
-                .ok_or(MetadataError::NumericalOverflowError)?;
-
-            // If this flag is set we are allowing any and all creators to be marked as verified
-            // without further checking.  This can only be done in special circumstances when the
-            // metadata is fully trusted such as when minting a limited edition.  Note we are still
-            // checking that creator share adds up to 100%.
-            if allow_direct_creator_writes {
-                continue;
-            }
-
-            // If this specific creator (of this loop iteration) is a signer and an update
-            // authority, then we are fine with this creator either setting or clearing its
-            // own `creator.verified` flag.
-            if update_authority_is_signer && **address == *update_authority {
-                continue;
-            }
-
-            // If the previous two conditions are not true then we check the state in the existing
-            // metadata creators array (if it exists) before allowing `creator.verified` to be set.
-            if let Some(existing_creators_map) = &existing_creators_map {
-                if existing_creators_map.contains_key(address) {
-                    // If this specific creator (of this loop iteration) is in the existing
-                    // creator's array, then it's `creator.verified` flag must match the existing
-                    // state.
-                    if creator.verified && !existing_creators_map[address].verified {
-                        return Err(MetadataError::CannotVerifyAnotherCreator.into());
-                    } else if !creator.verified && existing_creators_map[address].verified {
-                        return Err(MetadataError::CannotUnverifyAnotherCreator.into());
-                    }
-                } else if creator.verified {
-                    // If this specific creator is not in the existing creator's array, then we
-                    // cannot set `creator.verified`.
-                    return Err(MetadataError::CannotVerifyAnotherCreator.into());
-                }
-            } else if creator.verified {
-                // If there is no existing creators array, we cannot set `creator.verified`.
-                return Err(MetadataError::CannotVerifyAnotherCreator.into());
-            }
-        }
-
-        // Ensure share total is 100%.
-        if share_total != 100 {
-            return Err(MetadataError::ShareTotalMustBe100.into());
-        }
-
-        // Next make sure there were not any existing creators that were already verified but not
-        // listed in the new creator's array.
-        if allow_direct_creator_writes {
-            return Ok(());
-        } else if let Some(existing_creators_map) = &existing_creators_map {
-            for (address, existing_creator) in existing_creators_map {
-                // If this specific existing creator (of this loop iteration is a signer and an
-                // update authority, then we are fine with this creator clearing its own
-                // `creator.verified` flag.
-                if update_authority_is_signer && **address == *update_authority {
-                    continue;
-                } else if !new_creators_map.contains_key(address) && existing_creator.verified {
-                    return Err(MetadataError::CannotUnverifyAnotherCreator.into());
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
-
-/// assert initialized account
-pub fn assert_initialized<T: Pack + IsInitialized>(
-    account_info: &AccountInfo,
-) -> Result<T, ProgramError> {
-    let account: T = T::unpack_unchecked(&account_info.data.borrow())?;
-    if !account.is_initialized() {
-        Err(MetadataError::Uninitialized.into())
-    } else {
-        Ok(account)
-    }
-}
-
-/// Create account almost from scratch, lifted from
-/// <https://github.com/solana-labs/solana-program-library/tree/master/associated-token-account/program/src/processor.rs#L51-L98>
-#[inline(always)]
-pub fn create_or_allocate_account_raw<'a>(
-    program_id: Pubkey,
-    new_account_info: &AccountInfo<'a>,
-    system_program_info: &AccountInfo<'a>,
-    payer_info: &AccountInfo<'a>,
-    size: usize,
-    signer_seeds: &[&[u8]],
-) -> ProgramResult {
-    let rent = &Rent::get()?;
-    let required_lamports = rent
-        .minimum_balance(size)
-        .max(1)
-        .saturating_sub(new_account_info.lamports());
-
-    if required_lamports > 0 {
-        msg!("Transfer {} lamports to the new account", required_lamports);
-        invoke(
-            &system_instruction::transfer(payer_info.key, new_account_info.key, required_lamports),
-            &[
-                payer_info.clone(),
-                new_account_info.clone(),
-                system_program_info.clone(),
-            ],
-        )?;
-    }
-
-    let accounts = &[new_account_info.clone(), system_program_info.clone()];
-
-    msg!("Allocate space for the account");
-    invoke_signed(
-        &system_instruction::allocate(new_account_info.key, size.try_into().unwrap()),
-        accounts,
-        &[signer_seeds],
-    )?;
-
-    msg!("Assign the account to the owning program");
-    invoke_signed(
-        &system_instruction::assign(new_account_info.key, &program_id),
-        accounts,
-        &[signer_seeds],
-    )?;
-
-    Ok(())
-}
-
-/// Resize an account using realloc, lifted from Solana Cookbook
-#[inline(always)]
-pub fn resize_or_reallocate_account_raw<'a>(
-    target_account: &AccountInfo<'a>,
-    funding_account: &AccountInfo<'a>,
-    system_program: &AccountInfo<'a>,
-    new_size: usize,
-) -> ProgramResult {
-    let rent = Rent::get()?;
-    let new_minimum_balance = rent.minimum_balance(new_size);
-
-    let lamports_diff = new_minimum_balance.saturating_sub(target_account.lamports());
-    invoke(
-        &system_instruction::transfer(funding_account.key, target_account.key, lamports_diff),
-        &[
-            funding_account.clone(),
-            target_account.clone(),
-            system_program.clone(),
-        ],
-    )?;
-
-    target_account.realloc(new_size, false)?;
-
-    Ok(())
-}
-
-/// Close src_account and transfer lamports to dst_account, lifted from Solana Cookbook
-#[inline(always)]
-pub fn close_account_raw<'a>(
-    dest_account_info: &AccountInfo<'a>,
-    src_account_info: &AccountInfo<'a>,
-) -> ProgramResult {
-    let dest_starting_lamports = dest_account_info.lamports();
-    **dest_account_info.lamports.borrow_mut() = dest_starting_lamports
-        .checked_add(src_account_info.lamports())
-        .unwrap();
-    **src_account_info.lamports.borrow_mut() = 0;
-
-    let mut src_data = src_account_info.data.borrow_mut();
-    src_data.fill(0);
-
-    Ok(())
-}
-
-pub fn assert_update_authority_is_correct(
-    metadata: &Metadata,
-    update_authority_info: &AccountInfo,
-) -> ProgramResult {
-    if metadata.update_authority != *update_authority_info.key {
-        return Err(MetadataError::UpdateAuthorityIncorrect.into());
-    }
-
-    if !update_authority_info.is_signer {
-        return Err(MetadataError::UpdateAuthorityIsNotSigner.into());
-    }
-
-    Ok(())
-}
-
-/// Unpacks COption from a slice, taken from token program
-fn unpack_coption_key(src: &[u8; 36]) -> Result<COption<Pubkey>, ProgramError> {
-    let (tag, body) = array_refs![src, 4, 32];
-    match *tag {
-        [0, 0, 0, 0] => Ok(COption::None),
-        [1, 0, 0, 0] => Ok(COption::Some(Pubkey::new_from_array(*body))),
-        _ => Err(ProgramError::InvalidAccountData),
-    }
-}
-
-/// Cheap method to just grab owner Pubkey from token account, instead of deserializing entire thing
-pub fn get_owner_from_token_account(
-    token_account_info: &AccountInfo,
-) -> Result<Pubkey, ProgramError> {
-    // TokeAccount layout:   mint(32), owner(32), ...
-    let data = token_account_info.try_borrow_data()?;
-    let owner_data = array_ref![data, 32, 32];
-    Ok(Pubkey::new_from_array(*owner_data))
-}
-
-pub fn get_mint_authority(account_info: &AccountInfo) -> Result<COption<Pubkey>, ProgramError> {
-    // In token program, 36, 8, 1, 1 is the layout, where the first 36 is mint_authority
-    // so we start at 0.
-    let data = account_info.try_borrow_data().unwrap();
-    let authority_bytes = array_ref![data, 0, 36];
-
-    unpack_coption_key(authority_bytes)
-}
-
-pub fn get_mint_freeze_authority(
-    account_info: &AccountInfo,
-) -> Result<COption<Pubkey>, ProgramError> {
-    let data = account_info.try_borrow_data().unwrap();
-    let authority_bytes = array_ref![data, 36 + 8 + 1 + 1, 36];
-
-    unpack_coption_key(authority_bytes)
-}
-
-/// cheap method to just get supply off a mint without unpacking whole object
-pub fn get_mint_supply(account_info: &AccountInfo) -> Result<u64, ProgramError> {
-    // In token program, 36, 8, 1, 1 is the layout, where the first 8 is supply u64.
-    // so we start at 36.
-    let data = account_info.try_borrow_data()?;
-
-    // If we don't check this and an empty account is passed in, we get a panic when
-    // the array_ref! macro tries to index into the data.
-    if data.is_empty() {
-        return Err(ProgramError::InvalidAccountData);
-    }
-
-    let bytes = array_ref![data, 36, 8];
-
-    Ok(u64::from_le_bytes(*bytes))
-}
-
-/// cheap method to just get supply off a mint without unpacking whole object
-pub fn get_mint_decimals(account_info: &AccountInfo) -> Result<u8, ProgramError> {
-    // In token program, 36, 8, 1, 1, is the layout, where the first 1 is decimals u8.
-    // so we start at 36.
-    let data = account_info.try_borrow_data()?;
-
-    // If we don't check this and an empty account is passed in, we get a panic when
-    // we try to index into the data.
-    if data.is_empty() {
-        return Err(ProgramError::InvalidAccountData);
-    }
-
-    Ok(data[44])
-}
-
-pub fn assert_mint_authority_matches_mint(
-    mint_authority: &COption<Pubkey>,
-    mint_authority_info: &AccountInfo,
-) -> ProgramResult {
-    match mint_authority {
-        COption::None => {
-            return Err(MetadataError::InvalidMintAuthority.into());
-        }
-        COption::Some(key) => {
-            if mint_authority_info.key != key {
-                return Err(MetadataError::InvalidMintAuthority.into());
-            }
-        }
-    }
-
-    if !mint_authority_info.is_signer {
-        return Err(MetadataError::NotMintAuthority.into());
-    }
-
-    Ok(())
-}
-
-pub fn assert_supply_invariance(
-    master_edition: &MasterEditionV1,
-    printing_mint: &Mint,
-    new_supply: u64,
-) -> ProgramResult {
-    // The supply of printed tokens and the supply of the master edition should, when added, never exceed max supply.
-    // Every time a printed token is burned, master edition.supply goes up by 1.
-    if let Some(max_supply) = master_edition.max_supply {
-        let current_supply = printing_mint
-            .supply
-            .checked_add(master_edition.supply)
-            .ok_or(MetadataError::NumericalOverflowError)?;
-        let new_proposed_supply = current_supply
-            .checked_add(new_supply)
-            .ok_or(MetadataError::NumericalOverflowError)?;
-        if new_proposed_supply > max_supply {
-            return Err(MetadataError::PrintingWouldBreachMaximumSupply.into());
-        }
-    }
-
-    Ok(())
-}
+use crate::{
+    assertions::{
+        assert_derivation, assert_initialized, assert_mint_authority_matches_mint, assert_owned_by,
+        assert_token_program_matches_package,
+        collection::assert_collection_update_is_valid,
+        edition::{assert_edition_is_not_mint_authority, assert_edition_valid},
+        metadata::assert_data_valid,
+        uses::assert_valid_use,
+    },
+    deser::clean_write_metadata,
+    error::MetadataError,
+    state::{
+        get_reservation_list, CollectionDetails, DataV2, Edition, EditionMarker, Key,
+        MasterEditionV2, Metadata, TokenMetadataAccount, TokenStandard, Uses, EDITION,
+        EDITION_MARKER_BIT_SIZE, MAX_EDITION_LEN, MAX_EDITION_MARKER_SIZE, MAX_MASTER_EDITION_LEN,
+        MAX_METADATA_LEN, MAX_NAME_LENGTH, MAX_SYMBOL_LENGTH, MAX_URI_LENGTH, PREFIX,
+    },
+};
 
 pub fn transfer_mint_authority<'a>(
     edition_key: &Pubkey,
@@ -433,34 +78,6 @@ pub fn transfer_mint_authority<'a>(
         msg!("Finished setting freeze authority");
     } else {
         return Err(MetadataError::NoFreezeAuthoritySet.into());
-    }
-
-    Ok(())
-}
-
-pub fn assert_rent_exempt(rent: &Rent, account_info: &AccountInfo) -> ProgramResult {
-    if !rent.is_exempt(account_info.lamports(), account_info.data_len()) {
-        Err(MetadataError::NotRentExempt.into())
-    } else {
-        Ok(())
-    }
-}
-
-// Todo deprecate this for assert derivation
-pub fn assert_edition_valid(
-    program_id: &Pubkey,
-    mint: &Pubkey,
-    edition_account_info: &AccountInfo,
-) -> ProgramResult {
-    let edition_seeds = &[
-        PREFIX.as_bytes(),
-        program_id.as_ref(),
-        mint.as_ref(),
-        EDITION.as_bytes(),
-    ];
-    let (edition_key, _) = Pubkey::find_program_address(edition_seeds, program_id);
-    if edition_key != *edition_account_info.key {
-        return Err(MetadataError::InvalidEditionKey.into());
     }
 
     Ok(())
@@ -765,170 +382,6 @@ pub fn mint_limited_edition<'a>(
         mint_authority_info,
         token_program_account_info,
     )?;
-
-    Ok(())
-}
-
-pub fn spl_token_burn(params: TokenBurnParams<'_, '_>) -> ProgramResult {
-    let TokenBurnParams {
-        mint,
-        source,
-        authority,
-        token_program,
-        amount,
-        authority_signer_seeds,
-    } = params;
-    let mut seeds: Vec<&[&[u8]]> = vec![];
-    if let Some(seed) = authority_signer_seeds {
-        seeds.push(seed);
-    }
-    let result = invoke_signed(
-        &spl_token::instruction::burn(
-            token_program.key,
-            source.key,
-            mint.key,
-            authority.key,
-            &[authority.key],
-            amount,
-        )?,
-        &[source, mint, authority],
-        seeds.as_slice(),
-    );
-    result.map_err(|_| MetadataError::TokenBurnFailed.into())
-}
-
-pub fn spl_token_close(params: TokenCloseParams<'_, '_>) -> ProgramResult {
-    let TokenCloseParams {
-        account,
-        destination,
-        owner,
-        authority_signer_seeds,
-        token_program,
-    } = params;
-    let mut seeds: Vec<&[&[u8]]> = vec![];
-    if let Some(seed) = authority_signer_seeds {
-        seeds.push(seed);
-    }
-    let result = invoke_signed(
-        &spl_token::instruction::close_account(
-            token_program.key,
-            account.key,
-            destination.key,
-            owner.key,
-            &[],
-        )?,
-        &[account, destination, owner, token_program],
-        seeds.as_slice(),
-    );
-    result.map_err(|_| MetadataError::TokenCloseFailed.into())
-}
-
-/// TokenBurnParams
-pub struct TokenBurnParams<'a: 'b, 'b> {
-    /// mint
-    pub mint: AccountInfo<'a>,
-    /// source
-    pub source: AccountInfo<'a>,
-    /// amount
-    pub amount: u64,
-    /// authority
-    pub authority: AccountInfo<'a>,
-    /// authority_signer_seeds
-    pub authority_signer_seeds: Option<&'b [&'b [u8]]>,
-    /// token_program
-    pub token_program: AccountInfo<'a>,
-}
-
-/// TokenCloseParams
-pub struct TokenCloseParams<'a: 'b, 'b> {
-    /// Token account
-    pub account: AccountInfo<'a>,
-    /// Destination for redeemed SOL.
-    pub destination: AccountInfo<'a>,
-    /// Owner of the token account.
-    pub owner: AccountInfo<'a>,
-    /// authority_signer_seeds
-    pub authority_signer_seeds: Option<&'b [&'b [u8]]>,
-    /// token_program
-    pub token_program: AccountInfo<'a>,
-}
-
-pub fn spl_token_mint_to(params: TokenMintToParams<'_, '_>) -> ProgramResult {
-    let TokenMintToParams {
-        mint,
-        destination,
-        authority,
-        token_program,
-        amount,
-        authority_signer_seeds,
-    } = params;
-    let mut seeds: Vec<&[&[u8]]> = vec![];
-    if let Some(seed) = authority_signer_seeds {
-        seeds.push(seed);
-    }
-    let result = invoke_signed(
-        &spl_token::instruction::mint_to(
-            token_program.key,
-            mint.key,
-            destination.key,
-            authority.key,
-            &[],
-            amount,
-        )?,
-        &[mint, destination, authority, token_program],
-        seeds.as_slice(),
-    );
-    result.map_err(|_| MetadataError::TokenMintToFailed.into())
-}
-
-/// TokenMintToParams
-pub struct TokenMintToParams<'a: 'b, 'b> {
-    /// mint
-    pub mint: AccountInfo<'a>,
-    /// destination
-    pub destination: AccountInfo<'a>,
-    /// amount
-    pub amount: u64,
-    /// authority
-    pub authority: AccountInfo<'a>,
-    /// authority_signer_seeds
-    pub authority_signer_seeds: Option<&'b [&'b [u8]]>,
-    /// token_program
-    pub token_program: AccountInfo<'a>,
-}
-
-pub fn assert_derivation(
-    program_id: &Pubkey,
-    account: &AccountInfo,
-    path: &[&[u8]],
-) -> Result<u8, ProgramError> {
-    let (key, bump) = Pubkey::find_program_address(path, program_id);
-    if key != *account.key {
-        return Err(MetadataError::DerivedKeyInvalid.into());
-    }
-    Ok(bump)
-}
-
-pub fn assert_signer(account_info: &AccountInfo) -> ProgramResult {
-    if !account_info.is_signer {
-        Err(ProgramError::MissingRequiredSignature)
-    } else {
-        Ok(())
-    }
-}
-
-pub fn assert_owned_by(account: &AccountInfo, owner: &Pubkey) -> ProgramResult {
-    if account.owner != owner {
-        Err(MetadataError::IncorrectOwner.into())
-    } else {
-        Ok(())
-    }
-}
-
-pub fn assert_token_program_matches_package(token_program_info: &AccountInfo) -> ProgramResult {
-    if *token_program_info.key != spl_token::id() {
-        return Err(MetadataError::InvalidTokenProgram.into());
-    }
 
     Ok(())
 }
@@ -1284,83 +737,6 @@ pub fn process_mint_new_edition_from_master_edition_via_token_logic<'a>(
     )?;
     Ok(())
 }
-pub fn assert_currently_holding(
-    program_id: &Pubkey,
-    owner_info: &AccountInfo,
-    metadata_info: &AccountInfo,
-    metadata: &Metadata,
-    mint_info: &AccountInfo,
-    token_account_info: &AccountInfo,
-) -> ProgramResult {
-    assert_owned_by(metadata_info, program_id)?;
-    assert_owned_by(mint_info, &spl_token::id())?;
-
-    let token_account: Account = assert_initialized(token_account_info)?;
-
-    assert_owned_by(token_account_info, &spl_token::id())?;
-
-    if token_account.owner != *owner_info.key {
-        return Err(MetadataError::InvalidOwner.into());
-    }
-
-    if token_account.mint != *mint_info.key {
-        return Err(MetadataError::MintMismatch.into());
-    }
-
-    if token_account.amount < 1 {
-        return Err(MetadataError::NotEnoughTokens.into());
-    }
-
-    if token_account.mint != metadata.mint {
-        return Err(MetadataError::MintMismatch.into());
-    }
-    Ok(())
-}
-
-pub fn assert_freeze_authority_matches_mint(
-    freeze_authority: &COption<Pubkey>,
-    freeze_authority_info: &AccountInfo,
-) -> ProgramResult {
-    match freeze_authority {
-        COption::None => {
-            return Err(MetadataError::InvalidFreezeAuthority.into());
-        }
-        COption::Some(key) => {
-            if freeze_authority_info.key != key {
-                return Err(MetadataError::InvalidFreezeAuthority.into());
-            }
-        }
-    }
-    Ok(())
-}
-
-pub fn assert_delegated_tokens(
-    delegate: &AccountInfo,
-    mint_info: &AccountInfo,
-    token_account_info: &AccountInfo,
-) -> ProgramResult {
-    assert_owned_by(mint_info, &spl_token::id())?;
-
-    let token_account: Account = assert_initialized(token_account_info)?;
-
-    assert_owned_by(token_account_info, &spl_token::id())?;
-
-    if token_account.mint != *mint_info.key {
-        return Err(MetadataError::MintMismatch.into());
-    }
-
-    if token_account.amount < 1 {
-        return Err(MetadataError::NotEnoughTokens.into());
-    }
-
-    if token_account.delegate == COption::None
-        || token_account.delegated_amount != token_account.amount
-        || token_account.delegate.unwrap() != *delegate.key
-    {
-        return Err(MetadataError::InvalidDelegate.into());
-    }
-    Ok(())
-}
 
 pub fn increment_collection_size(
     metadata: &mut Metadata,
@@ -1405,24 +781,6 @@ pub fn decrement_collection_size(
         msg!("No collection details found. Cannot decrement collection size.");
         Err(MetadataError::UnsizedCollection.into())
     }
-}
-
-pub fn assert_verified_member_of_collection(
-    item_metadata: &Metadata,
-    collection_metadata: &Metadata,
-) -> ProgramResult {
-    if let Some(ref collection) = item_metadata.collection {
-        if collection_metadata.mint != collection.key {
-            return Err(MetadataError::NotAMemberOfCollection.into());
-        }
-        if !collection.verified {
-            return Err(MetadataError::NotVerifiedMemberOfCollection.into());
-        }
-    } else {
-        return Err(MetadataError::NotAMemberOfCollection.into());
-    }
-
-    Ok(())
 }
 
 pub fn check_token_standard(
@@ -1471,16 +829,4 @@ pub fn is_print_edition(
     let is_correct_type = Edition::from_account_info(edition_account_info).is_ok();
 
     is_correct_type && mint_decimals == 0 && mint_supply == 1
-}
-
-pub fn assert_edition_is_not_mint_authority(mint_account_info: &AccountInfo) -> ProgramResult {
-    let mint = Mint::unpack_from_slice(*mint_account_info.try_borrow_mut_data()?)?;
-
-    let (edition_pda, _) = find_master_edition_account(mint_account_info.key);
-
-    if mint.mint_authority == COption::Some(edition_pda) {
-        return Err(MetadataError::MissingEditionAccount.into());
-    }
-
-    Ok(())
 }
